xserver xkeyboard-config mesa git update Jan 19 2018
xserver          commit a5e9bcad7ad0887f804905b482894b85751519fb
xkeyboard-config commit 0418c71e56ba2675cf1daa939fa87bf079f0fdae
mesa             commit 8ff8c82630ccef75c13c84b5b32b45dda976f4ec

+++ mesalib/src/mesa/main/multisample.c
@@ -87,7 +87,7 @@ _mesa_GetMultisamplefv(GLenum pname, GLuint index, GLfloat * val)
 
    switch (pname) {
    case GL_SAMPLE_POSITION: {
-      if ((int) index >= ctx->DrawBuffer->Visual.samples) {
+      if (index >= ctx->DrawBuffer->Visual.samples) {
          _mesa_error( ctx, GL_INVALID_VALUE, "glGetMultisamplefv(index)" );
          return;
       }
